# Optimize passing multiple paths to usort CLI